refactor: Checklist files refactor
Changed ordering and posistioning of checklist component files
