+++ src/components/Checklist/ChecklistNotes/index.js
@@ -1,0 +1,2 @@
+export { default as ChecklistNotes } from './ChecklistNotes';
+export { default as ChecklistNotesModal } from './ChecklistNotesModal';
